requirements.txt
nose
yanc
pyhamcrest
mock
mockito
